+++ CHANGELOG.md
@@ -110,6 +110,7 @@ to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 - `drop` event
 - `itemdurability` event, condition and variable
 - `numbercompare` condition
+- `cancelOnLogout` argument for `folder` event
 - Things that are also added in 1.12.X:
     - new line support for `journal_lore` in `messages.yml`
     - FastAsyncWorldEdit compatibility